Generalized deploy to S3/Cloudfront
Cloudfront caches our stuff so we don't have as high of S3 costs, but
the tradeoff is we need to invalidate the cache after a deploy if we
want our content to update in a timely fashion.
Add methods to both:
a) list all the files that we put in place as part of the deploy
b) invoke aws cli tools to invalidate the caches
And automate the process of deploying/updating.

I'm removing the psd file here because the space in its filename was a
problem, and it doesn't seem relevant to fullaudit (it's not referenced
anywhere).