Delete existing built _site directory before building
We've had a few errant doubleposts show up, and I think it's partly
because people's _site directories had old content lying around.

This wipes out _site so we're starting fresh.

The True Fix for this is having CI, but "LPL Jenkins" is not a world I
want to be in yet.