Update Jekyll version to stay supported
This update has been deployed to staging and looks fine there.
You'll want to do a `bundle install` after merging this.

`I checked the targets of a bunch of URLs in staging's build (as per the
update caveats) and it all seems to have worked.

Built with:
```
✦ ❯ ruby --version
ruby 2.5.5p157 (2019-03-15 revision 67260) [x86_64-linux-gnu]
```

This is the result of a `bundle update`, because we haven't updated it
in about a year and the old sass library had become unsupported.

I'm also deploying it to prod so that we can get an updated warrant
canary.